Logig move to fragment
