--- AwesomeCalculator/app/src/main/java/com/example/drcreeper/awesomecalculator/fragments/CalkuulatorFragment.java
@@ -1,19 +1,0 @@
-package com.example.drcreeper.awesomecalculator.fragments;
-
-import android.app.Fragment;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.example.drcreeper.awesomecalculator.R;
-
-public class CalkuulatorFragment extends Fragment {
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.calculator_fragment,container,false);
-    }
-
-}
