+++ AwesomeCalculator/app/src/main/res/layout-land/activity_main.xml
